pyproject.toml
[tool.poetry]
name = "pictag"
version = "0.1.0"
description = ""
authors = ["AmaneTsu <amanetsukasa724@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
Pillow = "^9.0.1"
easy-thumbnails = "^2.8.1"
django-admin-decorators = "^0.1"
django-debug-toolbar = "^3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
